ResourceEvaluation entities extend an abstract class
